# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FerreMax CRM/ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: lowStock
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/{id}/adjust-stock:
    post:
      operationId: adjustProductStock
      tags: [products]
      summary: Adjust product stock
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockAdjustment"
      responses:
        "200":
          description: Stock adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  # ─── SALES ───────────────────────────────────────────────────────────────────
...
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: List quotes
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Quotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Create quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}:
    get:
      operationId: getQuote
      tags: [quotes]
      summary: Get quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
    patch:
      operationId: updateQuote
      tags: [quotes]
      summary: Update quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}/convert:
    post:
      operationId: convertQuoteToSale
      tags: [quotes]
      summary: Convert quote to sale
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Sale created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"

  # ─── SUPPLIERS ───────────────────────────────────────────────────────────────
...
  /purchases:
    get:
      operationId: listPurchases
      tags: [purchases]
      summary: List purchase orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: supplierId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Purchases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"
    post:
      operationId: createPurchase
      tags: [purchases]
      summary: Create purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
  /purchases/{id}:
    get:
      operationId: getPurchase
      tags: [purchases]
      summary: Get purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
    patch:
      operationId: updatePurchase
      tags: [purchases]
      summary: Update purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"

  # ─── WHATSAPP ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, sku, salePrice, stock, minStock, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: string
        barcode:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        purchasePrice:
          type: number
        salePrice:
          type: number
        wholesalePrice:
          type: ["number", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        brand:
          type: ["string", "null"]
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        stock:
          type: integer
        minStock:
          type: integer
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, sku, salePrice, stock, minStock]
      properties:
        name:
          type: string
        sku:
          type: string
        barcode:
          type: string
        description:
          type: string
        purchasePrice:
          type: number
        salePrice:
          type: number
        wholesalePrice:
          type: number
        categoryId:
          type: integer
        brand:
          type: string
        supplierId:
          type: integer
        stock:
          type: integer
        minStock:
          type: integer
        imageUrl:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        sku:
          type: string
        barcode:
          type: string
        description:
          type: string
        purchasePrice:
          type: number
        salePrice:
          type: number
        wholesalePrice:
          type: number
        categoryId:
          type: integer
        brand:
          type: string
        supplierId:
          type: integer
        stock:
          type: integer
        minStock:
          type: integer
        imageUrl:
          type: string
    StockAdjustment:
      type: object
      required: [quantity, type, reason]
      properties:
        quantity:
          type: integer
        type:
          type: string
        reason:
          type: string

    # ── SALE ──────────────────────────────────────────────────────────────────
    SaleItem:
      type: object
      required: [productId, productName, quantity, unitPrice, total]
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        discount:
          type: ["number", "null"]
        total:
          type: number
    Sale:
      type: object
      required: [id, total, status, paymentMethod, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        subtotal:
          type: number
        discount:
          type: ["number", "null"]
        tax:
          type: ["number", "null"]
        total:
          type: number
        status:
          type: string
        paymentMethod:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Quote:
      type: object
      required: [id, total, status, validUntil, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        subtotal:
          type: number
        discount:
          type: ["number", "null"]
        tax:
          type: ["number", "null"]
        total:
          type: number
        status:
          type: string
        notes:
          type: ["string", "null"]
        validUntil:
          type: string
        createdAt:
          type: string
    QuoteInput:
      type: object
      required: [items, validUntil]
      properties:
        customerId:
          type: integer
        items:
          type: array
          items:
            type: object
            required: [productId, quantity, unitPrice]
            properties:
              productId:
                type: integer
              quantity:
                type: integer
              unitPrice:
                type: number
              discount:
                type: number
        discount:
          type: number
        tax:
          type: number
        notes:
          type: string
        validUntil:
          type: string
    QuoteUpdate:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string
        validUntil:
          type: string

    # ── SUPPLIER ──────────────────────────────────────────────────────────────
...
    Purchase:
      type: object
      required: [id, supplierId, supplierName, total, status, createdAt]
      properties:
        id:
          type: integer
        supplierId:
          type: integer
        supplierName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        total:
          type: number
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PurchaseInput:
      type: object
      required: [supplierId, items]
      properties:
        supplierId:
          type: integer
        items:
          type: array
          items:
            type: object
            required: [productId, quantity, unitPrice]
            properties:
              productId:
                type: integer
              quantity:
                type: integer
              unitPrice:
                type: number
        notes:
          type: string
    PurchaseUpdate:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string

    # ── WHATSAPP ──────────────────────────────────────────────────────────────
...